+++ viewer/src/models/cad/delegates.ts
@@ -2,14 +2,6 @@
  * Copyright 2020 Cognite AS
  */
 
-import { DetermineSectorsInput } from './types';
-import { WantedSector } from '../../data/model/WantedSector';
-
-/**
- * Delegate for determining what sectors to load for a given frustum.
- */
-export type DetermineSectorsDelegate = (params: DetermineSectorsInput) => WantedSector[];
-
 /**
  * Delegate that handles removal of a sector from a view (e.g. by removing it from the scene).
  */
