+++ viewer/src/models/cad/types.ts
@@ -5,19 +5,6 @@
 import { Box3 } from '../../utils/Box3';
 import { mat4, vec3 } from 'gl-matrix';
 import { PrimitiveAttributes } from '../../workers/types/parser.types';
-import { CadLoadingHints } from './CadLoadingHints';
-
-/**
- * Input to DetermineSectorsDelegate.
- */
-export interface DetermineSectorsInput {
-  readonly sectorScene: SectorScene;
-  readonly cameraFov: number;
-  readonly cameraPosition: vec3;
-  readonly cameraModelMatrix: mat4;
-  readonly projectionMatrix: mat4;
-  readonly loadingHints?: CadLoadingHints;
-}
 
 // TODO 2019-11-12 larsmoa: Move and rename to something general (not specific
 // for sector data).
